+++ CHANGELOG.md
@@ -9,6 +9,7 @@ https://keepachangelog.com/en/1.0.0/
 https://www.markdownguide.org/basic-syntax/
 
 * [Unreleased](#unreleased)
+* [0.8.1](#081---2024-03-07)
 * [0.8.0](#080---2023-11-20)
 * [0.7.2](#072---2023-07-03)
 * [0.7.1](#071---2023-07-02)
@@ -31,6 +32,29 @@ https://www.markdownguide.org/basic-syntax/
 
 ## [Unreleased]
 
+## [0.8.1] - 2024-03-07
+
+### Added
+
+- Add functionality that create foreign key instead of constraint that check if foreign key belongs to tenant [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added property createForeignKeyConstraintWithTenantColumn in com.github.starnowski.posmulten.configuration.core.model.SharedSchemaContextConfiguration type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added com.github.starnowski.posmulten.configuration.yaml.model.PrimaryKeyDefinition.NameForFunctionThatChecksIfRecordExistsInTableNotBlank type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added property createForeignKeyConstraintWithTenantColumn in com.github.starnowski.posmulten.configuration.yaml.model.SharedSchemaContextConfiguration type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added getter and setter method for property createForeignKeyConstraintWithTenantColumn in com.github.starnowski.posmulten.configuration.yaml.model.SharedSchemaContextConfiguration type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added com.github.starnowski.posmulten.configuration.yaml.validation.groups.NameForFunctionThatChecksIfRecordExistsInTableNotBlankGroupResolver type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added com.github.starnowski.posmulten.configuration.yaml.validation.groups.ValidatorGroupResolver type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added com.github.starnowski.posmulten.configuration.yaml.validation.groups.ValidatorGroupsResolver type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added com.github.starnowski.posmulten.postgresql.core.DefaultForeignKeyConstraintStatementParameters type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added com.github.starnowski.posmulten.postgresql.core.ForeignKeyConstraintStatementProducer type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added com.github.starnowski.posmulten.postgresql.core.IForeignKeyConstraintStatementParameters type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added com.github.starnowski.posmulten.postgresql.core.context.DefaultSharedSchemaContextBuilder#setIgnoreCreationOfConstraintThatChecksIfRecordBelongsToCurrentTenant(Boolean) method. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added com.github.starnowski.posmulten.postgresql.core.context.DefaultSharedSchemaContextBuilder#setCreateForeignKeyConstraintWithTenantColumn(Boolean) method. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added property createForeignKeyConstraintWithTenantColumn in com.github.starnowski.posmulten.postgresql.core.context.SharedSchemaContextRequest type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added property ignoreCreationOfConstraintThatChecksIfRecordBelongsToCurrentTenant in com.github.starnowski.posmulten.postgresql.core.context.SharedSchemaContextRequest type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added getter and setter method for property createForeignKeyConstraintWithTenantColumn in com.github.starnowski.posmulten.postgresql.core.context.SharedSchemaContextRequest type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added getter and setter method for property ignoreCreationOfConstraintThatChecksIfRecordBelongsToCurrentTenant in com.github.starnowski.posmulten.postgresql.core.context.SharedSchemaContextRequest type. [301](https://github.com/starnowski/posmulten/issues/301)
+    - Added com.github.starnowski.posmulten.postgresql.core.context.enrichers.ForeignKeyConstraintSQLDefinitionsEnricher type. [301](https://github.com/starnowski/posmulten/issues/301)
+    
 ## [0.8.0] - 2023-11-20
 
 ### Added
@@ -74,7 +98,7 @@ https://www.markdownguide.org/basic-syntax/
 
 ### Added
 
-- Added dectorator for sqlDefinitions that replace template value like {{schema}} to any passed value in all definitions [283](https://github.com/starnowski/posmulten/issues/283)
+- Added decorator for sqlDefinitions that replace template value like {{schema}} to any passed value in all definitions [283](https://github.com/starnowski/posmulten/issues/283)
     - Added com.github.starnowski.posmulten.configuration.jar.Constants type. [283](https://github.com/starnowski/posmulten/issues/283)
     - Added com.github.starnowski.posmulten.configuration.jar.DDLScriptsGeneratorRunner type. [283](https://github.com/starnowski/posmulten/issues/283)
     - Added com.github.starnowski.posmulten.configuration.jar.SystemPropertiesDefaultDecoratorContextSupplier type. [283](https://github.com/starnowski/posmulten/issues/283)